S2EExecutor: moved memory debugging code to a separate file
